MNT: separate different positioner types
