+++ tests/test_positioner.py
@@ -6,9 +6,9 @@ import unittest
 # from copy import copy
 
 import epics
-from ophyd.controls.positioner import (Positioner, PVPositioner, EpicsMotor)
-from ophyd.controls.signal import (EpicsSignal, EpicsSignalRO)
-from ophyd.controls.device import (Component as C)
+from ophyd.controls import (Positioner, PVPositioner, EpicsMotor)
+from ophyd.controls import (EpicsSignal, EpicsSignalRO)
+from ophyd.controls import (Component as C)
 
 logger = logging.getLogger(__name__)
